Refactor: moved stuff to a fresh admincontroller
