+++ FoodApp/Controllers/AdminController.cs
@@ -1,0 +1,75 @@
+﻿using FoodApp.Data;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Vereyon.Web;
+
+namespace FoodApp.Controllers
+{
+    public class AdminController : Controller
+
+    {
+        private readonly ApplicationDbContext _context;
+        private readonly IFlashMessage _flashMessage;
+        public AdminController(ApplicationDbContext context, IFlashMessage flashMessage)
+        {
+            _context = context;
+            _flashMessage = flashMessage;
+
+
+        }
+        [Authorize(Roles = "Administrator")]
+        public async Task<IActionResult> RequestList()
+        {
+            return View("RequestList", await _context.Restaurant.Where(r => r.State == Models.Restaurant.RestaurantState.Pending).ToListAsync());
+        }
+        [Authorize(Roles = "Administrator")]
+        public async Task<IActionResult> Evaluate(string id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var restaurant = await _context.Restaurant
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (restaurant == null)
+            {
+                return NotFound();
+            }
+
+            return View(restaurant);
+        }
+        [Authorize(Roles = "Administrator")]
+        public async Task<IActionResult> Confirm(string id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+            var restaurant = await _context.Restaurant
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (restaurant == null)
+            {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+
+                restaurant.State = Models.Restaurant.RestaurantState.Confirmed;
+                _context.Update(restaurant);
+                _flashMessage.Confirmation($"{restaurant.Name} confirmed successfully!");
+                await _context.SaveChangesAsync();
+
+                return RedirectToAction(nameof(RequestList));
+            }
+            _flashMessage.Danger($"Something went wrong successfully! :(");
+            return RedirectToAction(nameof(RequestList));
+        }
+    }
+}
