+++ FoodApp/Controllers/RestaurantsController.cs
@@ -35,55 +35,7 @@ namespace FoodApp.Controllers
             _distanceProcessor = new DistanceProcessor();
 
         }
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult> RequestList()
-        {
-            return View("RequestList",await _context.Restaurant.Where(r => r.State==Models.Restaurant.RestaurantState.Pending) .ToListAsync());
-        }
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult> Evaluate(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var restaurant = await _context.Restaurant
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (restaurant == null)
-            {
-                return NotFound();
-            }
-
-            return View(restaurant);
-        }
-        [Authorize(Roles = "Administrator")]
-        public async Task<IActionResult> Confirm(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-            var restaurant = await _context.Restaurant
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (restaurant == null)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-
-                    restaurant.State = Models.Restaurant.RestaurantState.Confirmed;
-                    _context.Update(restaurant);
-                    _flashMessage.Confirmation($"{restaurant.Name} confirmed successfully!");
-                    await _context.SaveChangesAsync();
-                
-                return RedirectToAction(nameof(RequestList));
-            }
-            _flashMessage.Danger($"Something went wrong successfully! :(");
-            return RedirectToAction(nameof(RequestList));
-        }
+        
         //restaurants requests stuff above
         //guest stuff below
         /////////////////////////////////////////////////////////
